+++ CHANGELOG.md
@@ -12,6 +12,8 @@
   - Tile
   - New
 * Added support for glob inputs and a variable number of output images.
+* Moved user guide into separate file, replaced --generate-readme option
+  with --generate-guide
 
 # 0.1.0 (2019-09-15)